+++ lib/chef-dk/command/generator_commands/attribute.rb
@@ -15,14 +15,23 @@
 # limitations under the License.
 #
 
-source 'https://rubygems.org'
+require 'chef-dk/command/generator_commands/cookbook_code_file'
 
-gemspec :name => "chef-dk"
+module ChefDK
+  module Command
+    module GeneratorCommands
+      # chef generate attribute [path/to/cookbook_root] NAME
+      class Attribute < CookbookCodeFile
 
-gem "chef", ">= 11.14.0.rc.2"
+        banner "Usage: chef generate attribute [path/to/cookbook] NAME [options]"
 
-group(:dev) do
-  gem 'guard'
-  gem 'guard-rspec'
-  gem 'ruby_gntp'
+        options.merge!(SharedGeneratorOptions.options)
+
+        def recipe
+          'attribute'
+        end
+      end
+    end
+  end
 end
+
